Refactor filtering out duplicate patterns by name
